requirements.txt
asgiref==3.7.2
boto3==1.34.44
botocore==1.34.44
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.7
django-extra-views==0.14.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-storages==1.14.2
google-api-core==2.17.1
google-auth==2.28.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
jmespath==1.0.1
Pillow==10.1.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3

